requirements.txt
firebase-admin
customtkinter
tk
Pillow
requests